+++ upp-dto/src/main/java/com/upp/dto/generate/OnlineorderinfoExample.java
@@ -1,5 +1,6 @@
 package com.upp.dto.generate;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Iterator;
@@ -1762,52 +1763,52 @@ public class OnlineorderinfoExample {
             return (Criteria) this;
         }
 
-        public Criteria andTransamtEqualTo(Double value) {
+        public Criteria andTransamtEqualTo(BigDecimal value) {
             addCriterion("TRANSAMT =", value, "transamt");
             return (Criteria) this;
         }
 
-        public Criteria andTransamtNotEqualTo(Double value) {
+        public Criteria andTransamtNotEqualTo(BigDecimal value) {
             addCriterion("TRANSAMT <>", value, "transamt");
             return (Criteria) this;
         }
 
-        public Criteria andTransamtGreaterThan(Double value) {
+        public Criteria andTransamtGreaterThan(BigDecimal value) {
             addCriterion("TRANSAMT >", value, "transamt");
             return (Criteria) this;
         }
 
-        public Criteria andTransamtGreaterThanOrEqualTo(Double value) {
+        public Criteria andTransamtGreaterThanOrEqualTo(BigDecimal value) {
             addCriterion("TRANSAMT >=", value, "transamt");
             return (Criteria) this;
         }
 
-        public Criteria andTransamtLessThan(Double value) {
+        public Criteria andTransamtLessThan(BigDecimal value) {
             addCriterion("TRANSAMT <", value, "transamt");
             return (Criteria) this;
         }
 
-        public Criteria andTransamtLessThanOrEqualTo(Double value) {
+        public Criteria andTransamtLessThanOrEqualTo(BigDecimal value) {
             addCriterion("TRANSAMT <=", value, "transamt");
             return (Criteria) this;
         }
 
-        public Criteria andTransamtIn(List<Double> values) {
+        public Criteria andTransamtIn(List<BigDecimal> values) {
             addCriterion("TRANSAMT in", values, "transamt");
             return (Criteria) this;
         }
 
-        public Criteria andTransamtNotIn(List<Double> values) {
+        public Criteria andTransamtNotIn(List<BigDecimal> values) {
             addCriterion("TRANSAMT not in", values, "transamt");
             return (Criteria) this;
         }
 
-        public Criteria andTransamtBetween(Double value1, Double value2) {
+        public Criteria andTransamtBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("TRANSAMT between", value1, value2, "transamt");
             return (Criteria) this;
         }
 
-        public Criteria andTransamtNotBetween(Double value1, Double value2) {
+        public Criteria andTransamtNotBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("TRANSAMT not between", value1, value2, "transamt");
             return (Criteria) this;
         }
@@ -1822,52 +1823,52 @@ public class OnlineorderinfoExample {
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtEqualTo(Double value) {
+        public Criteria andRefundedamtEqualTo(BigDecimal value) {
             addCriterion("REFUNDEDAMT =", value, "refundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtNotEqualTo(Double value) {
+        public Criteria andRefundedamtNotEqualTo(BigDecimal value) {
             addCriterion("REFUNDEDAMT <>", value, "refundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtGreaterThan(Double value) {
+        public Criteria andRefundedamtGreaterThan(BigDecimal value) {
             addCriterion("REFUNDEDAMT >", value, "refundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtGreaterThanOrEqualTo(Double value) {
+        public Criteria andRefundedamtGreaterThanOrEqualTo(BigDecimal value) {
             addCriterion("REFUNDEDAMT >=", value, "refundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtLessThan(Double value) {
+        public Criteria andRefundedamtLessThan(BigDecimal value) {
             addCriterion("REFUNDEDAMT <", value, "refundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtLessThanOrEqualTo(Double value) {
+        public Criteria andRefundedamtLessThanOrEqualTo(BigDecimal value) {
             addCriterion("REFUNDEDAMT <=", value, "refundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtIn(List<Double> values) {
+        public Criteria andRefundedamtIn(List<BigDecimal> values) {
             addCriterion("REFUNDEDAMT in", values, "refundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtNotIn(List<Double> values) {
+        public Criteria andRefundedamtNotIn(List<BigDecimal> values) {
             addCriterion("REFUNDEDAMT not in", values, "refundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtBetween(Double value1, Double value2) {
+        public Criteria andRefundedamtBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("REFUNDEDAMT between", value1, value2, "refundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andRefundedamtNotBetween(Double value1, Double value2) {
+        public Criteria andRefundedamtNotBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("REFUNDEDAMT not between", value1, value2, "refundedamt");
             return (Criteria) this;
         }
@@ -1882,52 +1883,52 @@ public class OnlineorderinfoExample {
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtEqualTo(Double value) {
+        public Criteria andUnrefundedamtEqualTo(BigDecimal value) {
             addCriterion("UNREFUNDEDAMT =", value, "unrefundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtNotEqualTo(Double value) {
+        public Criteria andUnrefundedamtNotEqualTo(BigDecimal value) {
             addCriterion("UNREFUNDEDAMT <>", value, "unrefundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtGreaterThan(Double value) {
+        public Criteria andUnrefundedamtGreaterThan(BigDecimal value) {
             addCriterion("UNREFUNDEDAMT >", value, "unrefundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtGreaterThanOrEqualTo(Double value) {
+        public Criteria andUnrefundedamtGreaterThanOrEqualTo(BigDecimal value) {
             addCriterion("UNREFUNDEDAMT >=", value, "unrefundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtLessThan(Double value) {
+        public Criteria andUnrefundedamtLessThan(BigDecimal value) {
             addCriterion("UNREFUNDEDAMT <", value, "unrefundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtLessThanOrEqualTo(Double value) {
+        public Criteria andUnrefundedamtLessThanOrEqualTo(BigDecimal value) {
             addCriterion("UNREFUNDEDAMT <=", value, "unrefundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtIn(List<Double> values) {
+        public Criteria andUnrefundedamtIn(List<BigDecimal> values) {
             addCriterion("UNREFUNDEDAMT in", values, "unrefundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtNotIn(List<Double> values) {
+        public Criteria andUnrefundedamtNotIn(List<BigDecimal> values) {
             addCriterion("UNREFUNDEDAMT not in", values, "unrefundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtBetween(Double value1, Double value2) {
+        public Criteria andUnrefundedamtBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("UNREFUNDEDAMT between", value1, value2, "unrefundedamt");
             return (Criteria) this;
         }
 
-        public Criteria andUnrefundedamtNotBetween(Double value1, Double value2) {
+        public Criteria andUnrefundedamtNotBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("UNREFUNDEDAMT not between", value1, value2, "unrefundedamt");
             return (Criteria) this;
         }
@@ -2082,52 +2083,52 @@ public class OnlineorderinfoExample {
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtEqualTo(Double value) {
+        public Criteria andFeeamtEqualTo(BigDecimal value) {
             addCriterion("FEEAMT =", value, "feeamt");
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtNotEqualTo(Double value) {
+        public Criteria andFeeamtNotEqualTo(BigDecimal value) {
             addCriterion("FEEAMT <>", value, "feeamt");
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtGreaterThan(Double value) {
+        public Criteria andFeeamtGreaterThan(BigDecimal value) {
             addCriterion("FEEAMT >", value, "feeamt");
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtGreaterThanOrEqualTo(Double value) {
+        public Criteria andFeeamtGreaterThanOrEqualTo(BigDecimal value) {
             addCriterion("FEEAMT >=", value, "feeamt");
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtLessThan(Double value) {
+        public Criteria andFeeamtLessThan(BigDecimal value) {
             addCriterion("FEEAMT <", value, "feeamt");
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtLessThanOrEqualTo(Double value) {
+        public Criteria andFeeamtLessThanOrEqualTo(BigDecimal value) {
             addCriterion("FEEAMT <=", value, "feeamt");
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtIn(List<Double> values) {
+        public Criteria andFeeamtIn(List<BigDecimal> values) {
             addCriterion("FEEAMT in", values, "feeamt");
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtNotIn(List<Double> values) {
+        public Criteria andFeeamtNotIn(List<BigDecimal> values) {
             addCriterion("FEEAMT not in", values, "feeamt");
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtBetween(Double value1, Double value2) {
+        public Criteria andFeeamtBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("FEEAMT between", value1, value2, "feeamt");
             return (Criteria) this;
         }
 
-        public Criteria andFeeamtNotBetween(Double value1, Double value2) {
+        public Criteria andFeeamtNotBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("FEEAMT not between", value1, value2, "feeamt");
             return (Criteria) this;
         }
